Rust: Restructure the project
