--- src/bindings/rust/elektra/src/elektra_kdb/mod.rs
@@ -1,7 +1,0 @@
-mod error;
-mod kdb;
-
-pub use self::error::{
-    KDBError, KDBErrorWrapper, LogicalError, PermanentError, ResourceError, ValidationError,
-};
-pub use self::kdb::KDB;
