--- src/bindings/rust/elektra/src/elektra_key/mod.rs
@@ -1,13 +1,0 @@
-mod builder;
-mod error;
-mod key;
-mod readable;
-mod readonly;
-mod writable;
-
-pub use self::builder::KeyBuilder;
-pub use self::error::KeyError;
-pub use self::key::{BinaryKey, StringKey};
-pub use self::readable::ReadableKey;
-pub use self::readonly::ReadOnly;
-pub use self::writable::WriteableKey;
