--- src/bindings/rust/elektra/src/elektra_keyset/mod.rs
@@ -1,7 +1,0 @@
-mod error;
-mod keyset;
-mod iterator;
-
-pub use self::error::KeySetError;
-pub use self::keyset::{KeySet,Cursor};
-pub use self::iterator::{ReadOnlyStringKeyIter,StringKeyIter};
