+++ src/bindings/rust/elektra/src/iterator.rs
