+++ src/bindings/rust/elektra/src/writable.rs
